flint::commands::gen_preview

Struct GenPreviewApp

struct GenPreviewApp {

    spec: ProcGenSpec,
    spec_path: PathBuf,
    registry: GeneratorRegistry,
    field_specs: Vec<ParamFieldSpec>,
    output_kind: OutputKind,
    reload_flag: Arc<Mutex<bool>>,
    _watcher: Option<Debouncer<RecommendedWatcher>>,
    params: Map<String, Value>,
    dirty: bool,
    last_change: Instant,
    debounce_ms: u64,
    current_seed: u64,
    last_skeleton: Option<SkeletonDef>,
    last_mesh: Option<MeshData>,
    last_images: Vec<ImageData>,
    selected_image_tab: usize,
    texture_zoom: f32,
    texture_offset: [f32; 2],
    texture_tiled: bool,
    texture_tile_count: usize,
    gen_time_ms: f64,
    last_world: Option<FlintWorld>,
    args: GenPreviewArgs,
    window: Option<Arc<Window>>,
    render_context: Option<RenderContext>,
    scene_renderer: Option<SceneRenderer>,
    camera: Camera,
    orbit: OrbitCameraController,
    last_frame_time: Instant,
    egui_ctx: Context,
    egui_winit: Option<State>,
    egui_renderer: Option<Renderer>,
    egui_textures: Vec<(String, TextureHandle)>,
    show_ui: bool,
    frame_times: VecDeque<Instant>,
    fps: f32,
    last_fps_update: Instant,
    lod_count: usize,
    selected_lod: usize,
    watcher_paused_until: Option<Instant>,
    status_message: Option<(String, Instant)>,

}

Fields
spec: ProcGenSpecspec_path: PathBufregistry: GeneratorRegistryfield_specs: Vec<ParamFieldSpec>output_kind: OutputKindreload_flag: Arc<Mutex<bool>>_watcher: Option<Debouncer<RecommendedWatcher>>params: Map<String, Value>dirty: boollast_change: Instantdebounce_ms: u64current_seed: u64last_skeleton: Option<SkeletonDef>last_mesh: Option<MeshData>last_images: Vec<ImageData>selected_image_tab: usizetexture_zoom: f32texture_offset: [f32; 2]texture_tiled: booltexture_tile_count: usizegen_time_ms: f64last_world: Option<FlintWorld>args: GenPreviewArgswindow: Option<Arc<Window>>render_context: Option<RenderContext>scene_renderer: Option<SceneRenderer>camera: Cameraorbit: OrbitCameraControllerlast_frame_time: Instantegui_ctx: Contextegui_winit: Option<State>egui_renderer: Option<Renderer>egui_textures: Vec<(String, TextureHandle)>show_ui: boolframe_times: VecDeque<Instant>fps: f32last_fps_update: Instantlod_count: usizeselected_lod: usizewatcher_paused_until: Option<Instant>status_message: Option<(String, Instant)>
Implementations

impl GenPreviewApp

fn initialize(&mut self, event_loop: &ActiveEventLoop)

fn regenerate(&mut self)

fn upload_mesh(&mut self, mesh: &MeshData, _lod: usize)

Convert procgen MeshData to render vertices and upload to the GPU.

fn upload_submeshes(&mut self, mesh: &MeshData)

Upload one procedural mesh per submesh, each with the correct material color.

fn auto_fit_to_mesh(&mut self, mesh: &MeshData)

fn update_skeleton_overlay(&mut self)

Update skeleton overlay from stored skeleton data.

fn update_egui_textures(&mut self)

Convert ImageData to egui texture handles for display.

fn check_file_reload(&mut self)

fn mark_dirty(&mut self)

fn should_regenerate(&self) -> bool

fn save_spec(&mut self)

fn export_output(&mut self)

fn export_mesh_to_path(&mut self, path: &Path)

fn export_image_to_path(&mut self, img: &ImageData, path: &Path)

fn save_spec_as(&mut self)

fn set_status(&mut self, msg: &str)

fn render_egui(&mut self, target_view: &TextureView)

fn reload_from_disk(&mut self)