flint::commands::tex_edit

Struct TexEditApp

struct TexEditApp {

    spec: ProcGenSpec,
    spec_path: PathBuf,
    map_params: TextureMapParams,
    reload_flag: Arc<Mutex<bool>>,
    _watcher: Option<Debouncer<RecommendedWatcher>>,
    snarl: Snarl<TextureNode>,
    selected_node: Option<NodeId>,
    dirty: bool,
    last_change: Instant,
    debounce_ms: u64,
    current_seed: u64,
    last_images: Vec<ImageData>,
    snapshots: Vec<TextureField>,
    gen_time_ms: f64,
    node_preview_textures: HashMap<NodeId, TextureHandle>,
    selected_image_tab: usize,
    selected_channel: Option<String>,
    selected_snapshot_node: Option<usize>,
    texture_zoom: f32,
    texture_offset: [f32; 2],
    texture_tiled: bool,
    args: TexEditArgs,
    window: Option<Arc<Window>>,
    render_context: Option<RenderContext>,
    egui_ctx: Context,
    egui_winit: Option<State>,
    egui_renderer: Option<Renderer>,
    egui_textures: Vec<(String, TextureHandle)>,
    show_ui: bool,
    frame_times: VecDeque<Instant>,
    fps: f32,
    last_fps_update: Instant,
    undo_stack: Vec<UndoSnapshot>,
    redo_stack: Vec<UndoSnapshot>,
    watcher_paused_until: Option<Instant>,
    status_message: Option<(String, Instant)>,

}

Fields
spec: ProcGenSpecspec_path: PathBufmap_params: TextureMapParamsreload_flag: Arc<Mutex<bool>>_watcher: Option<Debouncer<RecommendedWatcher>>snarl: Snarl<TextureNode>selected_node: Option<NodeId>dirty: boollast_change: Instantdebounce_ms: u64current_seed: u64last_images: Vec<ImageData>snapshots: Vec<TextureField>gen_time_ms: f64node_preview_textures: HashMap<NodeId, TextureHandle>selected_image_tab: usizeselected_channel: Option<String>selected_snapshot_node: Option<usize>texture_zoom: f32texture_offset: [f32; 2]texture_tiled: boolargs: TexEditArgswindow: Option<Arc<Window>>render_context: Option<RenderContext>egui_ctx: Contextegui_winit: Option<State>egui_renderer: Option<Renderer>egui_textures: Vec<(String, TextureHandle)>show_ui: boolframe_times: VecDeque<Instant>fps: f32last_fps_update: Instantundo_stack: Vec<UndoSnapshot>redo_stack: Vec<UndoSnapshot>watcher_paused_until: Option<Instant>status_message: Option<(String, Instant)>
Implementations

impl TexEditApp

fn initialize(&mut self, event_loop: &ActiveEventLoop)

fn auto_assign_id(existing: &HashSet<String>, op_type: &str) -> String

Generate a unique node ID for an op type, given a set of existing IDs.

Returns the op_type itself if unused, otherwise appends _2, _3, etc.

fn ensure_all_ids(&mut self)

Ensure all Op nodes in the snarl have IDs assigned.

fn push_undo(&mut self)

fn undo(&mut self)

fn redo(&mut self)

fn collect_ops_toml(&self) -> Vec<Value>

fn restore_from_ops(&mut self, ops: &[Value])

fn mark_dirty(&mut self)

fn regenerate(&mut self)

fn update_node_previews(&mut self)

fn update_textures(&mut self)

fn save_spec(&mut self)

fn export_textures(&mut self)

fn reload_from_disk(&mut self)

fn ordered_op_nodes(&self) -> Vec<(NodeId, &TextureNode)>

Get the ordered list of op nodes (excluding Output) for pipeline execution.

impl TexEditApp

fn render_frame(&mut self)

fn draw_ui(&mut self)

fn draw_properties_panel(&mut self, ui: &mut Ui)

fn draw_preview_panel(&mut self, ui: &mut Ui)

fn draw_graph_panel(&mut self, ui: &mut Ui)